### 2023-03-20 #5285 @editorialbot
[PRE REVIEW]: PyVBMC: Efficient Bayesian inference in Python
**Submitting author:** @Bobby-Huggins (<a href="http://orcid.org/0009-0006-3475-5964">Bobby Huggins</a>)
**Repository:** https://github.com/acerbilab/pyvbmc
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.0
