### 2023-04-25 #5285 @arfon
I'm delighted to say that @rkurchin has kindly agreed to edit this submission for JOSS. Thanks again @rkurchin!
