### 2023-04-26 #5285 @rkurchin
👋 @tbrown122387, @ejhigson, @SuperKam91, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
