+++ JOSS/PyVBMC_JOSS.bib
@@ -95,7 +95,7 @@ volume = {153},
 pages = {108046},
 year = {2021},
 issn = {0306-4549},
-doi = {https://doi.org/10.1016/j.anucene.2020.108046},
+doi = {10.1016/j.anucene.2020.108046},
 url = {https://www.sciencedirect.com/science/article/pii/S0306454920307428},
 author = {Yifeng Che and Xu Wu and Giovanni Pastore and Wei Li and Koroush Shirvan},
 keywords = {Doped fuel, Variational Bayesian Monte Carlo (VBMC), Bayesian inference, Kriging, Principal Component Analysis (PCA)},
@@ -215,7 +215,7 @@ number = {3},
 pages = {245-260},
 year = {1991},
 issn = {0378-3758},
-doi = {https://doi.org/10.1016/0378-3758(91)90002-V},
+doi = {10.1016/0378-3758(91)90002-V},
 url = {https://www.sciencedirect.com/science/article/pii/037837589190002V},
 author = {A. O'Hagan},
 keywords = {Bayesian quadrature, numerical integration, Gaussian process, product rule, Gaussian quadrature},
