### 2023-04-28 #5285 @Bobby-Huggins
@rkurchin got it, the invalid DOI's should be fixed with the latest push.
