### 2023-04-28 #5285 @matt-graham
I'd be very happy to review this. Just for full disclosure Luigi Acerbi and I were on the same PhD program (in different years) at University of Edinburgh. As far as I can tell this doesn't constitute a conflict of interest under the guidelines at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html but just wanted to make this explicit in case it would be a problem.
