### 2023-05-03 acerbilab/pyvbmc#135 @matt-graham
`test_minimize_adam_matyas_with_noise` intermittently failing
Raising issue as part of JOSS review openjournals/joss-reviews/issues/5428

When running the tests with `platform linux -- Python 3.11.0, pytest-7.3.1, pluggy-1.0.0` the test in `pyvbmc/testing/vbmc/test_minimize_adam.py::test_minimize_adam_matyas_with_noise` intermittently fails with an `AssertionError`, for example two failing cases were
```
>       assert np.all(np.abs(x) < 1.0)
E       AssertionError: assert False
E        +  where False = <function all at 0x7fd42c38c860>(array([1.43098764, 1.41719139]) < 1.0)
E        +    where <function all at 0x7fd42c38c860> = np.all
E        +    and   array([1.43098764, 1.41719139]) = <ufunc 'absolute'>(array([1.43098764, 1.41719139]))
E        +      where <ufunc 'absolute'> = np.abs

pyvbmc/testing/vbmc/test_minimize_adam.py:70: AssertionError
````
```
>       assert np.all(np.abs(x) < 1.0)
E       AssertionError: assert False
E        +  where False = <function all at 0x7f3cda609bc0>(array([1.08179515, 1.15884769]) < 1.0)
E        +    where <function all at 0x7f3cda609bc0> = np.all
E        +    and   array([1.08179515, 1.15884769]) = <ufunc 'absolute'>(array([-1.08179515, -1.15884769]))
E        +      where <ufunc 'absolute'> = np.abs

pyvbmc/testing/vbmc/test_minimize_adam.py:70: AssertionError
```
but on other runs the test passes. From the difference in the reported value of the `x` array between the runs in which fails occurred it looks like the test is non-deterministic, and looking at the code it seems that random variates are being generated from the module-level Numpy random number generator rather than a separately seeded `Generator` instance:

https://github.com/acerbilab/pyvbmc/blob/b7c6e381be75dd508ac9663825fd77ed8885816d/pyvbmc/testing/vbmc/test_minimize_adam.py#L54-L71

It looks like the threshold being tested for here might be a bit tight as the probability of the test failing seems to be non-negligible. More generally I would recommend using seeded random number generators in all tests to avoid intermittent failures like this and make it easier to reproduce and debug failing tests. You could for example create [a shared `rng` fixture](https://docs.pytest.org/en/latest/how-to/fixtures.html#scope-sharing-fixtures-across-classes-modules-packages-or-session) in a `conftest.py` file and [parametrize it with one or more seeds passed in as a command line argument](https://stackoverflow.com/a/42145604)

```Python
import pytest
import numpy

def pytest_addoption(parser):
    parser.addoption("--seed", type=int, nargs="*", default=[28375632879561])

def pytest_generate_tests(metafunc):
    if "seed" in metafunc.fixturenames:
        metafunc.parametrize("seed", metafunc.config.getoption("seed"))

@pytest.fixture(scope="function")
def rng(seed):
    return numpy.random.default_rng(seed)
```

The above test would then be changed to something like

```Python
 def test_minimize_adam_matyas_with_noise(rng):
     def f(x_):
         val = 0.26 * (x_[0] ** 2 + x_[1] ** 2) - 0.48 * x_[0] * x_[1]
         grad_1 = 0.52 * x_[0] - 0.48 * x_[1]
         grad_2 = 0.52 * x_[1] - 0.48 * x_[0]
         grad = np.array([grad_1, grad_2]) + rng.normal(
             scale=3, size=(2,)
         )
         return val, grad
     ...
```

This would mean the default behaviour is to run with a single fixed seed but allow the option of running all tests using the `rng` fixture across multiple seeds by passing in values to the `--seed` option to the `pytest` command to check for robustness to different random draws.
