### 2023-05-03 acerbilab/pyvbmc#135 @Bobby-Huggins
Hi @matt-graham, thanks for opening the issue!

A lot of our tests are stochastic, but it seems this one has a narrow tolerance in particular. For this reason, we have `pytest-rerunfailures` in the dependencies and run e.g. our automated Workflow tests with `pytest --reruns=5`  to re-try any failed test up to 5 times (3 should be plenty already, 5 is extra-safe). We mention this in the developer instructions page of the docs, but since you bring it up we should probably add a note about it directed to end-users as well. Maybe even have `pytest` default to this behavior.

I do think we should look into the solution you propose, if for no other reason than to conform to best practices. I'm not certain at the moment how quickly we could migrate all the tests that have a stochastic component: There are some internal (non-test) changes we'd need to make too in order to have the exact behavior fixed by the `pytest` seed. In the meantime would you mind trying with (say) `--reruns=3`, just to confirm that it works as expected?
