### 2023-05-15 86769da
feat: keep track of wall-clock time in `vbmc.optimize()`.
