### 2023-05-16 acerbilab/pyvbmc#136 @matt-graham
Some possible mistakes in docstring for `VBMC` class
Raising issue as part of JOSS review openjournals/joss-reviews/issues/5428

The docstring for the `VBMC` class specifies the key in the `options` dict for specifying to use a noisy log target density as `specifytargetnoise`

https://github.com/acerbilab/pyvbmc/blob/b7c6e381be75dd508ac9663825fd77ed8885816d/pyvbmc/vbmc/vbmc.py#L51

while elsewhere in the documentation `specify_target_noise ` is used and this seem to correspond to what's assumed in the code

https://github.com/acerbilab/pyvbmc/blob/b7c6e381be75dd508ac9663825fd77ed8885816d/pyvbmc/vbmc/option_configs/basic_vbmc_options.ini#L19

There is also a note indicating PyVBMC does not currently support noisy target density evaluations which I think is no longer true (?) given there is a tutorial example for this case

https://github.com/acerbilab/pyvbmc/blob/b7c6e381be75dd508ac9663825fd77ed8885816d/pyvbmc/vbmc/vbmc.py#L114-L117
