### 2023-05-16 acerbilab/pyvbmc#136 @lacerbi
Thanks for spotting this!

- `specifytargetnoise` should be `specify_target_noise`;
- PyVBMC fully supports noisy target density evaluations; the note is a residue from an earlier alpha version and should be removed.

@Bobby-Huggins can you please fix this when you have the time?
