### 2023-05-16 acerbilab/pyvbmc#137 @matt-graham
`ValueError` raised in `FunctionLogger` when target density is zero
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5428

The code block

https://github.com/acerbilab/pyvbmc/blob/b7c6e381be75dd508ac9663825fd77ed8885816d/pyvbmc/function_logger/function_logger.py#L148-L156

raises a `ValueError` when the value returned by the function is non-scalar, non-finite or non-real. When the function corresponds to the logarithm of the target density, this excludes the target density taking value zero, and so it's logarithm being negative infinity. I imagine this might be a constraint implied by using a Gaussian process to fit to the log target density evaluations, so this check probably makes sense, but as it implies a (weak) constraint on the class of posterior distributions that can be targeted it may be worth explicitly mentioning this is disallowed in the documentation, and/or suggesting workarounds (possibly outputting a large in magnitude but finite negative value as a proxy for negative infinity?). I imagine some common cases where users might want to use a target density which may take exactly evaluate to zero are: (i) when the prior has support only on a subset of the specified domain but this doesn't correspond to a simple Cartesian product of intervals as covered by the lower / upper bound options, (ii) observation noise from some truncated distribution is assumed which will mean the likelihood may evaluate to zero, (iii) the model involves a numerical solver which may fail to converge for some choices of the parameters, for which a common resolution would be to output a target density of zero to exclude these regions of parameter space from posterior (I hit against this in the latter case when trying out PyVBMC with a model which involves solving a system of ordinary differential equations for which the solver fails to converge for some parameter settings).

As a sidenote, I might be wrong but I think the `np.any` call in the condition expression to the `if` block is not needed. `np.any` would usually be used on an array of boolean values. Here if `f_val_orig` is not a scalar and so `not np.isscalar(f_val_orig)` evaluates to `True` then because of [the short-circuit evaluation semantics of `or`](https://docs.python.org/3/reference/expressions.html#boolean-operations) then the whole expression will evaluate to `True` without evaluating the `not np.isfinite(f_val_orig)` and `not np.isreal(f_val_orig)` parts of the expression. If instead `f_val_orig` is a scalar then both `np.isfinite(f_val_orig)` and `not np.isfinite(f_val_orig)` will output scalars I believe and the overall expression will therefore always evaluate to either (scalar) `True` or `False` without the need for `np.any`.
