### 2023-05-16 acerbilab/pyvbmc#135 @matt-graham
> Hi @matt-graham, thanks for opening the issue!
>
> A lot of our tests are stochastic, but it seems this one has a narrow tolerance in particular. For this reason, we have `pytest-rerunfailures` in the dependencies and run e.g. our automated Workflow tests with `pytest --reruns=5` to re-try any failed test up to 5 times (3 should be plenty already, 5 is extra-safe). We mention this in the developer instructions page of the docs, but since you bring it up we should probably add a note about it directed to end-users as well. Maybe even have `pytest` default to this behavior.
>
> I do think we should look into the solution you propose, if for no other reason than to conform to best practices. I'm not certain at the moment how quickly we could migrate all the tests that have a stochastic component: There are some internal (non-test) changes we'd need to make too in order to have the exact behavior fixed by the `pytest` seed. In the meantime would you mind trying with (say) `--reruns=3`, just to confirm that it works as expected?

Thanks for the explanation @Bobby-Huggins, I had indeed missed the note about this in the developer documentation.

All tests do pass when rerunning with `--reruns=3`. If you are able to make the test behaviour deterministic though I would say it would be worth doing so as otherwise it can be difficult to be sure if an intermittently failing test is due to non-determinism from the use of an unseeded random number generator or some other cause.
