### 2023-05-17 acerbilab/pyvbmc#139 @matt-graham
Some options appear to not have any effect
Raising issue as part of JOSS review https://github.com/openjournals/joss-reviews/issues/5428

Some of the options documented at https://acerbilab.github.io/pyvbmc/api/options/vbmc_options.html do not appear to have any corresponding active code paths which use them. From some limited scanning, those that I've come across which this appears to apply to are:

  * `retry_max_fun_evals`
  * `optim_toolbox`
  * `repeated_acq_discount` - this appears to have been used at one point but the relevant code block is now commented out https://github.com/acerbilab/pyvbmc/blob/b7c6e381be75dd508ac9663825fd77ed8885816d/pyvbmc/vbmc/active_sample.py#L465
