### 2023-05-18 acerbilab/pyvbmc#137 @lacerbi
Thanks for the comment! This is actually a non-trivial issue to deal with when working with surrogate models of the *log*-density.
The original VBMC algorithm does not deal with `Inf`s and `NaN`s, assuming the output of the log-density is always well-behaved. This is indeed a limitation, for now we should probably mention it more explicitly in the documentation (besides the FAQ), as you also suggest above:
> it may be worth explicitly mentioning this is disallowed in the documentation

As for future plans, we do want to address the issues of zero-density (or `NaN`s)

A naive workaround of outputting an arbitrary large negative number will break the GP surrogate, but a heuristic dynamical approach that sets this value based on the current observed values might work, combined with (fake) observation noise that informs (Py)VBMC not to take that observation too seriously (for the latter, see [this preprint](https://arxiv.org/abs/2303.05263), Section 3.5 on noise shaping).

In an ideal world, we should use a logistic non-Gaussian likelihood for the GP model that informs the GP surrogate that the observation is *below a threshold*. However, introduction of non-Gaussian likelihoods requires approximate inference for the GP, which is a major modification of the algorithm and a separate research project in itself.
