### 2023-05-18 1d72471
docs: VBMC docstrings (issue 136).
