### 2023-05-18 d75983e
refactor: remove unnecessary code + options re: issues 137 + 138.
