+++ pyvbmc/vbmc/option_configs/basic_vbmc_options.ini
@@ -11,8 +11,6 @@ max_fun_evals = 50*(2+D)
 fun_evals_per_iter = 5
 # Required stable fcn evals for termination
 tol_stable_count = 60
-# Max number of target fcn evals on retry (0 = no retry)
-retry_max_fun_evals = 0
 # Number of variational components to refine posterior at termination
 min_final_components = 50
 # Target log joint function returns noise estimate (SD) as second output
