### 2023-05-18 bd6548b
Minor fixes from JoSS review + static linting (#140)
* chore: update pre-commit packages.

* fix: string_to_acq is not imported.

* docs: VBMC docstrings (issue 136).

* docs: update docs re: issues 137 and 135.

* refactor: remove unnecessary code + options re: issues 137 + 138.

---------

Co-authored-by: pipme <sjtulck@gmail.com>
