+++ pyvbmc/vbmc/option_configs/advanced_vbmc_options.ini
@@ -7,8 +7,6 @@ integer_vars = []
 noise_size = []
 # Max number of consecutive repeated measurements for noisy inputs
 max_repeated_observations = 0
-# Multiplicative discount True acquisition fcn to repeat measurement at the same location
-repeated_acq_discount = 1
 # Number of initial target fcn evals
 fun_eval_start = np.maximum(D, 10)
 # Base step size for stochastic gradient descent
@@ -33,8 +31,6 @@ output_fcn = []
 tol_stable_excpt_frac = 0.2
 # Evaluated fcn values at X0
 f_vals = []
-# Use Optimization Toolbox (if empty determine at runtime)
-optim_toolbox = []
 # Weighted proposal fcn for uncertainty search
 proposal_fcn = None
 # Automatic nonlinear rescaling of variables
