### 2023-05-19 acerbilab/pyvbmc#139 @Bobby-Huggins
Hi @matt-graham, thanks for pointing these out. As you've probably guessed, these options are from the MATLAB implementation but are unused in this implementation, at least for the time being. To avoid confusion we removed them just now in pull #140.
