### 2023-05-19 acerbilab/pyvbmc#136 @Bobby-Huggins
Fixed in pull #140.
