### 2023-05-19 acerbilab/pyvbmc#137 @Bobby-Huggins
Regarding the extraneous `np.any` call, you are correct @matt-graham: It has been removed in pull #140.
