### 2023-05-19 acerbilab/pyvbmc#137 @matt-graham
Thanks @Bobby-Huggins and for the additions to the documentation making the requirements for the target log density function more visible. Closing this as now resolved from my perspective.
