### 2023-05-23 #5428 @rkurchin
Some final steps here (see checklist above):
1. Does Mikko J. Arnos have an ORCID and if so, could you add it to the manuscript?
2. Please merge any and all changes from this review into your main branch and **issue a new version tag**. _(If you want to merge in the paper, you may, but it is not required that the actual manuscript live into the repo in perpetuity since JOSS will host it and you can simply add a badge link or whatever you like. But the actual changes to software and docs do need to be merged!)_
3. Create a DOI for the contents of the repo at the same commit corresponding to that version tag, e.g. using figshare or Zenodo. Please make sure that the metadata (version number, title, author list, etc.) match those of your manuscript.
4. Post a comment here with the version number and DOI.

I'll post some editorial feedback on the manuscript itself shortly.
