### 2023-05-23 #5428 @rkurchin
Editorial comments on manuscript (very nicely written, this is by far the fewest editorial comments I've ever given! 🤩):
- line 39: sample efficiency doesn't need to be hyphenated here
- 59: the bit starting from "For example:" is a sentence fragment. Could just join to previous sentence with commas, or perhaps break out into a bulleted list?
- 62: unfeasible -> infeasible

As a separate remark, this looks like really cool software and actually may be just the tool I need for some of my own work, so I'll likely start playing around with it myself! 😁
