### 2023-05-24 #5428 @Bobby-Huggins
Here is the DOI for the archive of the `v1.0.1` release (bumped from `v1.0.0` following the changes suggested during review): `10.5281/zenodo.7966315`
