+++ JOSS/PyVBMC_JOSS.bib
@@ -208,7 +208,7 @@ URL = {https://doi.org/10.1214/20-BA1200}
 }
 
 @article{ohagan_bayesian_1991,
-title = {Bayes–Hermite quadrature},
+title = {{Bayes–Hermite} quadrature},
 journal = {Journal of Statistical Planning and Inference},
 volume = {29},
 number = {3},
@@ -244,7 +244,7 @@ abstract = {Bayesian quadrature treats the problem of numerical integration as o
 ﻿@Article{harris_array_2020,
 author={Harris, Charles R.
 and Millman, K. Jarrod
-and van der Walt, St{\'e}fan J.
+and {van der Walt}, St{\'e}fan J.
 and Gommers, Ralf
 and Virtanen, Pauli
 and Cournapeau, David
@@ -255,10 +255,10 @@ and Smith, Nathaniel J.
 and Kern, Robert
 and Picus, Matti
 and Hoyer, Stephan
-and van Kerkwijk, Marten H.
+and {van Kerkwijk}, Marten H.
 and Brett, Matthew
 and Haldane, Allan
-and del R{\'i}o, Jaime Fern{\'a}ndez
+and {del R{\'i}o}, Jaime Fern{\'a}ndez
 and Wiebe, Mark
 and Peterson, Pearu
 and G{\'e}rard-Marchant, Pierre
