+++ docsrc/source/development.rst
@@ -11,11 +11,11 @@ The documentation is available at: https://acerbilab.github.io/pyvbmc/
 Installation instructions for developers
 ########################################
 
-Release versions of PyVBMC are available via ``pip`` (and soon, ``conda-forge``), but developers will need to work with the latest source code. They should follow these steps to install:
+Release versions of PyVBMC are available via ``pip`` and ``conda-forge``, but developers will need to work with the latest source code. They should follow these steps to install:
 
-1. (Optional, but recommended for development): Create a new environment in Conda and activate it. Requires Python 3.9 or newer::
+1. (Optional, but recommended for development): Create a new environment in Conda and activate it.
 
-     conda create --name pyvbmc-env python=3.9
+     conda create --name pyvbmc-env
      conda activate pyvbmc-env
 
 2. Clone the PyVBMC and GPyReg GitHub repos locally::
@@ -60,13 +60,13 @@ The code is formatted using `Black <https://pypi.org/project/black/>`_ with a li
 
 After installation, when you try to commit the staged files, git will automatically check the files and modify them for meeting the requirements of the hooks in ``.pre-commit-config.yaml``. The settings of the hooks are specified in ``pyproject.toml``. You need to restage the file if it gets modified by the hooks.
 
-If you want, you can also check with Pylint for more detailed errors and warnings (although Pylint seems to raise many false positives).
+If you want, you can also check with `ruff <https://beta.ruff.rs/docs/>` or `Pylint <https://www.pylint.org/>` for more detailed errors, warnings, and suggestions.
 
 Docstrings
 ----------
 
 The docstrings are generated following the `NumPy format <https://numpydoc.readthedocs.io/en/latest/format.html>`_.
-There are add-ons to generate docstring blueprints using IDE's.
+There are add-ons to generate docstring blueprints using IDEs.
 
 - See an example for a correct docstring from NumPy `here <https://numpydoc.readthedocs.io/en/latest/example.html>`__.
 - In PyVBMC, the ``VariationalPosterior`` class can be taken as an example of (mostly) correct docstring structure, see :mainbranch:`here <variational_posterior/variational_posterior.py>`.
@@ -120,15 +120,14 @@ Please keep the documentation up to date. (Sphinx logs possible issues when comp
 Exceptions
 ----------
 
-Currently, the aim is to use the standard Python exceptions whenever it is sensible.
-Here is a list of those `exceptions <https://docs.python.org/3/library/exceptions.html>`_.
+Please use standard Python exceptions whenever it is sensible. Here is a list of those `exceptions <https://docs.python.org/3/library/exceptions.html>`_.
 
 ``git`` commits
 ---------------
 
 Commits follow the `conventional commits <https://www.conventionalcommits.org/en/v1.0.0/>`_ style. This makes it easier to collaborate on the project. A cheat sheet is can be found `here <https://cheatography.com/albelop/cheat-sheets/conventional-commits/>`__
 
-Please do not submit pull requests with unfinished code or code which does not pass all tests. Work on feature branches whenever possible and sensible. All PR's must be approved by another developer before being merged to the main branch. `Read this <https://martinfowler.com/bliki/FeatureBranch.html>`_ ::
+Please do not submit pull requests with unfinished code or code which does not pass all tests. Work on feature branches whenever possible and sensible. All PRs must be approved by another developer before being merged to the main branch. `Read this <https://martinfowler.com/bliki/FeatureBranch.html>`_ ::
 
     git checkout -b <new-feature>
     [... do stuff and commit ...]
@@ -140,7 +139,7 @@ If you switch to an existing branch using ``git checkout``, remember to ``pull``
 Modules and code organization
 -----------------------------
 
[…]
