+++ _sources/development.rst
@@ -11,11 +11,11 @@ The documentation is available at: https://acerbilab.github.io/pyvbmc/
 Installation instructions for developers
 ########################################
 
-Release versions of PyVBMC are available via ``pip`` (and soon, ``conda-forge``), but developers will need to work with the latest source code. They should follow these steps to install:
+Release versions of PyVBMC are available via ``pip`` and ``conda-forge``, but developers will need to work with the latest source code. They should follow these steps to install:
 
-1. (Optional, but recommended for development): Create a new environment in Conda and activate it. Requires Python 3.9 or newer::
+1. (Optional, but recommended for development): Create a new environment in Conda and activate it.
 
-     conda create --name pyvbmc-env python=3.9
+     conda create --name pyvbmc-env
      conda activate pyvbmc-env
 
 2. Clone the PyVBMC and GPyReg GitHub repos locally::
@@ -37,22 +37,22 @@ Release versions of PyVBMC are available via ``pip`` (and soon, ``conda-forge``)
 
 We are using the dependencies listed in ``pyproject.toml``. Please list all used dependencies there. Dependencies are separated into basic dependencies, and optional development dependencies included under ``dev``.
 
-The necessary packages can be installed with `conda <https://docs.conda.io/projects/conda/en/latest/user-guide/install/>`_ or `pip <https://pypi.org/project/pip/>`_.
+The necessary packages can be installed with `conda <https://docs.conda.io/projects/conda/en/latest/user-guide/install/>`__ or `pip <https://pypi.org/project/pip/>`__.
 
 Coding conventions
 ##################
 
 We try to follow common conventions whenever possible. Some useful reading:
 
-- `PEP 8 -- Style Guide for Python Code <https://www.python.org/dev/peps/pep-0008/>`_
-- `Code style in The Hitchhiker's Guide to Python <https://docs.python-guide.org/writing/style/>`_
+- `PEP 8 -- Style Guide for Python Code <https://www.python.org/dev/peps/pep-0008/>`__
+- `Code style in The Hitchhiker's Guide to Python <https://docs.python-guide.org/writing/style/>`__
 
 These basic rules should be followed to ensure coherence and to make it easy for third parties to contribute. In the following, we list more detailed conventions. Please read carefully if you are contributing to PyVBMC.
 
 Code formatting
 ---------------
 
-The code is formatted using `Black <https://pypi.org/project/black/>`_ with a line length of 79, with the help of pre-commit hooks. To install and use::
+The code is formatted using `Black <https://pypi.org/project/black/>`__ with a line length of 79, with the help of pre-commit hooks. To install and use::
 
     pip install pre-commit
     pre-commit install
@@ -60,22 +60,22 @@ The code is formatted using `Black <https://pypi.org/project/black/>`_ with a li
 
 After installation, when you try to commit the staged files, git will automatically check the files and modify them for meeting the requirements of the hooks in ``.pre-commit-config.yaml``. The settings of the hooks are specified in ``pyproject.toml``. You need to restage the file if it gets modified by the hooks.
 
-If you want, you can also check with Pylint for more detailed errors and warnings (although Pylint seems to raise many false positives).
+If you want, you can also check with `ruff <https://beta.ruff.rs/docs/>`__ or `Pylint <https://www.pylint.org/>`__ for more detailed errors, warnings, and suggestions.
 
 Docstrings
 ----------
 
-The docstrings are generated following the `NumPy format <https://numpydoc.readthedocs.io/en/latest/format.html>`_.
-There are add-ons to generate docstring blueprints using IDE's.
+The docstrings are generated following the `NumPy format <https://numpydoc.readthedocs.io/en/latest/format.html>`__.
+There are add-ons to generate docstring blueprints using IDEs.
 
[…]
