+++ README.md
@@ -163,11 +163,11 @@ You may also want to check out [Bayesian Adaptive Direct Search in Python (PyBAD
     title = {PyVBMC: Efficient Bayesian inference in Python},
     author = {Bobby Huggins and Chengkun Li and Marlon Tobaben and Mikko J. Aarnos and Luigi Acerbi},
     publisher = {The Open Journal},
-    journal = {Journal of Open Source Software}
+    journal = {Journal of Open Source Software},
     url = {https://doi.org/10.21105/joss.05428},
     doi = {10.21105/joss.05428},
     year = {2023},
-    volume = {8}, number = {86}, pages = {5428},
+    volume = {8}, number = {86}, pages = {5428}
   }
 
 @article{acerbi2018variational,
